frontend-plugin-api: move out resolveInputOverrides
