+++ packages/frontend-plugin-api/src/wiring/createExtensionBlueprint.ts
@@ -19,7 +19,6 @@ import { Expand } from '../types';
 import {
   CreateExtensionOptions,
   ExtensionDefinition,
-  ResolvedExtensionInput,
   ResolvedExtensionInputs,
   VerifyExtensionFactoryOutput,
   createExtension,
@@ -28,13 +27,16 @@ import { z } from 'zod';
 import { ExtensionInput } from './createExtensionInput';
 import {
   AnyExtensionDataRef,
-  ExtensionDataRefToValue,
   ExtensionDataValue,
 } from './createExtensionDataRef';
 import {
   ExtensionDataContainer,
   createExtensionDataContainer,
 } from './createExtensionDataContainer';
+import {
+  ResolveInputValueOverrides,
+  resolveInputOverrides,
+} from './resolveInputOverrides';
 
 /**
  * @public
@@ -78,61 +80,6 @@ export type CreateExtensionBlueprintOptions<
 
   dataRefs?: TDataRefs;
 } & VerifyExtensionFactoryOutput<UOutput, UFactoryOutput>;
-
-/** @public */
-export type ResolveInputValueOverrides<
-  TInputs extends {
-    [inputName in string]: ExtensionInput<
-      AnyExtensionDataRef,
-      { optional: boolean; singleton: boolean }
-    >;
-  } = {
-    [inputName in string]: ExtensionInput<
-      AnyExtensionDataRef,
-      { optional: boolean; singleton: boolean }
-    >;
-  },
-> = Expand<
-  {
-    [KName in keyof TInputs as TInputs[KName] extends ExtensionInput<
-      any,
-      {
-        optional: infer IOptional extends boolean;
-        singleton: boolean;
-      }
-    >
-      ? IOptional extends true
-        ? never
-        : KName
-      : never]: TInputs[KName] extends ExtensionInput<
-      infer IDataRefs,
-      { optional: boolean; singleton: infer ISingleton extends boolean }
-    >
-      ? ISingleton extends true
-        ? Iterable<ExtensionDataRefToValue<IDataRefs>>
-        : Array<Iterable<ExtensionDataRefToValue<IDataRefs>>>
-      : never;
-  } & {
-    [KName in keyof TInputs as TInputs[KName] extends ExtensionInput<
-      any,
-      {
-        optional: infer IOptional extends boolean;
-        singleton: boolean;
-      }
-    >
-      ? IOptional extends true
-        ? KName
-        : never
-      : never]?: TInputs[KName] extends ExtensionInput<
-      infer IDataRefs,
-      { optional: boolean; singleton: infer ISingleton extends boolean }
-    >
-      ? ISingleton extends true
-        ? Iterable<ExtensionDataRefToValue<IDataRefs>>
-        : Array<Iterable<ExtensionDataRefToValue<IDataRefs>>>
-      : never;
-  }
->;
 
 /**
  * @public
@@ -252,84 +199,6 @@ export interface ExtensionBlueprint<
     string | undefined extends TNewNamespace ? TNamespace : TNewNamespace,
     string | undefined extends TNewName ? TName : TNewName
   >;
-}
-
-function expectArray<T>(value: T | T[]): T[] {
-  return value as T[];
-}
-function expectItem<T>(value: T | T[]): T {
-  return value as T;
-}
-
-/** @internal */
-export function resolveInputOverrides(
-  declaredInputs?: {
-    [inputName in string]: ExtensionInput<
-      AnyExtensionDataRef,
-      { optional: boolean; singleton: boolean }
-    >;
-  },
-  inputs?: {
-    [KName in string]?:
-      | ({ node: AppNode } & ExtensionDataContainer<any>)
-      | Array<{ node: AppNode } & ExtensionDataContainer<any>>;
-  },
-  inputOverrides?: ResolveInputValueOverrides,
-) {
-  if (!declaredInputs || !inputs || !inputOverrides) {
-    return inputs;
-  }
-
-  const newInputs: typeof inputs = {};
-  for (const name in declaredInputs) {
-    if (!Object.hasOwn(declaredInputs, name)) {
-      continue;
-    }
-    const declaredInput = declaredInputs[name];
-    const providedData = inputOverrides[name];
-    if (declaredInput.config.singleton) {
-      const originalInput = expectItem(inputs[name]);
-      if (providedData) {
-        const providedContainer = createExtensionDataContainer(
-          providedData as Iterable<ExtensionDataValue<any, any>>,
-          declaredInput.extensionData,
-        );
-        if (!originalInput) {
-          throw new Error(
-            `attempted to override data of input '${name}' but it is not present in the original inputs`,
-          );
-        }
-        newInputs[name] = Object.assign(providedContainer, {
-          name: (originalInput as ResolvedExtensionInput<any>).node,
-        }) as any;
-      }
-    } else {
-      const originalInput = expectArray(inputs[name]);
-      if (!Array.isArray(providedData)) {
-        throw new Error(
-          `override data provided for input '${name}' must be an array`,
-        );
-      }
-      if (
-        originalInput.length !== providedData.length &&
-        providedData.length > 0
-      ) {
-        throw new Error(
-          `override data provided for input '${name}' must match the length of the original inputs`,
-        );
-      }
-      newInputs[name] = providedData.map((data, i) => {
-        const providedContainer = createExtensionDataContainer(
-          data as Iterable<ExtensionDataValue<any, any>>,
-          declaredInput.extensionData,
-        );
-        return Object.assign(providedContainer, {
-          name: (originalInput[i] as ResolvedExtensionInput<any>).node,
-        }) as any;
-      });
-    }
-  }
-  return newInputs;
 }
 
 /**
